update policy
this new version makes the policy contract compatible with the new dollar.sol contract. specifically on line 156, the previous version of dollarPolicy.sol uses the now outdated totalDividend and totalDebt variables to estimate the circulating supply. This is now removed from dollars.sol in an effort to streamline the contract and only expose useful variables for consumption.